+++ boot/bootutil/CMakeLists.txt
@@ -22,6 +22,7 @@ target_sources(bootutil
         src/bootutil_img_security_cnt.c
         src/bootutil_misc.c
         src/bootutil_area.c
+        src/bootutil_loader.c
         src/bootutil_public.c
         src/caps.c
         src/encrypted.c
