+++ boot/espressif/CMakeLists.txt
@@ -242,6 +242,7 @@ set(bootutil_srcs
     ${BOOTUTIL_DIR}/src/bootutil_img_security_cnt.c
     ${BOOTUTIL_DIR}/src/bootutil_misc.c
     ${BOOTUTIL_DIR}/src/bootutil_area.c
+    ${BOOTUTIL_DIR}/src/bootutil_loader.c
     ${BOOTUTIL_DIR}/src/bootutil_public.c
     ${BOOTUTIL_DIR}/src/caps.c
     ${BOOTUTIL_DIR}/src/encrypted.c
